% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{augbd augmendted block design data using factors and subsamples}
\alias{augbd augmendted block design data using factors and subsamples}
\alias{augbd}
\title{trial data of an augmented block design experiment using factor with three levels. Beside it has subsample per genotype}
\format{data frame}
\source{
International Potato Center, potato and sweet potato experimental data.
}
\usage{
augbd
}
\description{
This dataset contains data which comes from experiment using factors and sub sampling.
}
\references{
This data is related to HiDAP fbcheck module.
}
