% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_calculations.R
\name{calculate_trait_variables}
\alias{calculate_trait_variables}
\title{Calculation of trait variables}
\usage{
calculate_trait_variables(fb, plot_size = NA, plant_den = NA, mgt, mtl,
  trial_type = NA)
}
\arguments{
\item{fb}{Data frame with the fieldbook}

\item{plot_size}{plot size}

\item{plant_den}{plant density}

\item{mgt}{Crop Managment Data}

\item{mtl}{Material List Data}

\item{trial_type}{the type of trial}
}
\description{
A generalized function to calculate all the posibles derivated trait from Fieldbook. Use as a 
backend sbformulas.
}

