% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitformat_excel.R
\name{col_validation}
\alias{col_validation}
\title{Generate conditional format for every trait column into fieldbook spreadsheet (Excel)}
\usage{
col_validation(file, fbsheet, trait, trait_dict)
}
\arguments{
\item{file}{The name of the file which contains your data fieldbook data. It must be an .xlsx file}

\item{fbsheet}{The name of the fieldbook sheet into the excel file.}

\item{trait}{The abbreviation of the trait used in fieldbooks.}

\item{trait_dict}{The trait dictionary on crop ontology format.
(paint with colours) the trait column to identify out of range values.}
}
\value{
An excel file with conditional format according trait conditions
}
\description{
This function use openxlsx package to give conditional format for fieldbook spreadsheet
}

