% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fctsubsample data with factor}
\alias{fctsubsample data with factor}
\alias{fctsubsample}
\title{trial data using factor with sub sampling at the same time.}
\format{data frame}
\source{
International Potato Center, potato and sweet potato experimental data.
}
\usage{
fctsubsample
}
\description{
This dataset contains data which comes from experiment using factors and sub sampling.
}
\references{
This data is related to HiDAP fbcheck module.
}
