% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.fb.param}
\alias{get.fb.param}
\title{Get fieldbook parameters into Excel Files}
\usage{
get.fb.param(file, sheet, parameter)
}
\arguments{
\item{file}{The file name}

\item{sheet}{The sheet name}

\item{parameter}{The column header to extract parameters.}
}
\description{
This function gets parameters or values from fieldbook excel file. Do an excel scrapping.
}

