% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.lb.control}
\alias{get.lb.control}
\title{Get the scale for late blight test to compare with other genotypes}
\usage{
get.lb.control(mtl)
}
\arguments{
\item{mtl}{The material list data}
}
\description{
Function calculate relative days for Late Blight Resistance.
}

