% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.rel.days}
\alias{get.rel.days}
\title{Get Relative Dates for Area Under the Curve(AUDPC)}
\usage{
get.rel.days(mgt)
}
\arguments{
\item{mgt}{The crop managemente data}
}
\description{
mgt function calculate relative days for Late Blight Resistance.
}

