% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_tools.R
\name{get_duplicate_db_id}
\alias{get_duplicate_db_id}
\title{Get duplicates of plot db id from fieldbook app files.}
\usage{
get_duplicate_db_id(data)
}
\arguments{
\item{data}{The name of the data frame.}
}
\value{
List of duplicated unique values and row position of plot ids
}
\description{
Get duplicates plot db id of fieldbook.
}
\author{
alonso medina
}
