% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fb_param}
\alias{get_fb_param}
\title{Get fieldbook parameters into Excel Files}
\usage{
get_fb_param(file, sheet, parameter)
}
\arguments{
\item{file}{The file name}

\item{sheet}{The sheet name}

\item{parameter}{The label of the factor we want to extract. Ex. "Experimental design", "Plot_Size"}
}
\description{
This function gets parameters or values from fieldbook excel file. Do an excel scrapping.
}

