% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sheet_data}
\alias{get_sheet_data}
\title{Get data from excel files}
\usage{
get_sheet_data(file, sheet)
}
\arguments{
\item{file}{The file name}

\item{sheet}{The sheet name}
}
\description{
This function gets all the data from excel files.
}
