% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_tools.R
\name{get_trait_crop}
\alias{get_trait_crop}
\title{Get the crop which one trait belong}
\usage{
get_trait_crop(trait, trait_dict, dsource = 1)
}
\arguments{
\item{trait}{trait abreviation}

\item{trait_dict}{trait dictionary}

\item{dsource}{source of the data \code{1}: Hidap,  \code{2}: FieldbookApp.}
}
\value{
The crop
}
\description{
Get the crop which one trait belong
}
\examples{
\dontrun{
# The data
data(potato_yield)
#str(potato_yield)
# NTP: Number of tuber planted
get_trait_crop("NTP", potato_yield)
}
}
\author{
omar benites
}
