% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_tools.R
\name{get_trait_name}
\alias{get_trait_name}
\title{Get the full name of the trait into crop trait dictionary}
\usage{
get_trait_name(trait, trait_dict, dsource = 1)
}
\arguments{
\item{trait}{trait abreviation}

\item{trait_dict}{trait dictionary}

\item{dsource}{source of the data \code{1}: Hidap,  \code{2}: FieldbookApp.}
}
\value{
The full name of the trait
}
\description{
Get the full name of the trait into crop trait dictionary
}
\examples{
\dontrun{
# The data
data(potato_yield)
#str(potato_yield)
# NTP: Number of tuber planted
get_trait_name("NTP", potato_yield)
}
}
\author{
omar benites
}
