% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_stats.R
\name{length2}
\alias{length2}
\title{Lenght of the values}
\usage{
length2(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector}

\item{na.rm}{Logical value to ommit NA values}
}
\description{
A lenght modified function to manipulate data in case of missing values.
}

