% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_excel_date}
\alias{mutate_excel_date}
\title{Get all the trait variables of fieldbooks, excluding all the fieldbook factors}
\usage{
mutate_excel_date(date_value)
}
\arguments{
\item{date_value}{A date to be fixed when you import dates from Excel}
}
\value{
A fixed date
}
\description{
When you import dates from Excel, it is quite common that some error or format issues happen. So,
this function changes from yyyy/mm/dd to ddd/mm/yyyy.
}
\examples{
\dontrun{
# The data
fix_excel_date(date_value="2013-05-22")
} 
}
\author{
omar benites
}
