% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_stats.R
\name{outlier_val}
\alias{outlier_val}
\title{Detect outlier trait values of fieldbook data}
\usage{
outlier_val(trait_data, f = 3)
}
\arguments{
\item{trait_data}{a vector which contains the data of the trait}

\item{f}{parameter for determining exteme values. By default it's 3.}
}
\value{
Return a list of two values ol (lower outlier bound) and lu (upper outlier bound)
}
\description{
this function calculates the outlier values using interquantile ranges to
determine atypical values in trait data.
}
\author{
Raul Eyzaguirre
}
