% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{post_sheet_data}
\alias{post_sheet_data}
\title{Post the fieldbook data into excel files}
\usage{
post_sheet_data(file, sheet, fieldbook)
}
\arguments{
\item{file}{The file name}

\item{sheet}{The sheet name}

\item{fieldbook}{The fieldbook data}
}
\description{
This function gets all the data from excel files.
}
