% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_crop_dictionary.R
\docType{data}
\name{potato_bulking}
\alias{potato_bulking}
\title{Trait dictionary for potato tuber bulking}
\format{A data frame with 39 columns and 53 rows.}
\source{
International Potato Center, potato experimental data.
}
\usage{
potato_bulking
}
\description{
This data set contains a trait dicionary based on Crop Ontology template for
Tuber Bulking trial in Potato. This trait dictionary is developed at International
Potato Center.
}
\references{
This data dictionary is related to CO_330
}

