% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_crop_dictionary.R
\docType{data}
\name{potato_yield}
\alias{potato_yield}
\title{Trait dictionary for potato yield trial}
\format{A data frame with 39 columns and 53 rows.}
\source{
International Potato Center, potato experimental data.
}
\usage{
potato_yield
}
\description{
This data set contains a trait dicionary based on Crop Ontology template for
Healthy Tuber and Yield trail in Potato. This trait dictionary is developed at International
Potato Center.
}
\references{
This data dictionary is related to CO_330
}
