% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reactive_excel_fb}
\alias{reactive_excel_fb}
\title{Read fieldbooks in reactive enviorment (Shiny)}
\usage{
reactive_excel_fb(file_id, sheet)
}
\arguments{
\item{file_id}{The id of the shinyInput for an fieldbook excel (input$shinyInput_id).}

\item{sheet}{The name of sheet which contains fieldbook data}
}
\value{
A data.frame contains the fieldbook data
}
\description{
Read fieldbooks in reactive enviorment (Shiny)
}
\author{
omar benites
}
