% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reactive_excel_metadata}
\alias{reactive_excel_metadata}
\title{Read metadata in reactive enviorment (Shiny)}
\usage{
reactive_excel_metadata(file_id, sheet)
}
\arguments{
\item{file_id}{The shinyInput's ID of the excel file (input$shinyInput_id).}

\item{sheet}{The name of sheet which contains meta data}
}
\value{
A data.frame contains the meta data
}
\description{
Read metadata in reactive enviorment (Shiny)
}
\author{
omar benites
}

