% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_tools.R
\name{render_categorical}
\alias{render_categorical}
\title{RJavascript generator for cualitative trait renders}
\usage{
render_categorical(scale_condition)
}
\arguments{
\item{scale_condition}{scale condition of the trait}
}
\description{
Function to generate Javascript code for quantitative trait validation in Rhansontable.
}
\author{
omar benites
}

