% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_tools.R
\name{render_dup_plotdbid_ext}
\alias{render_dup_plotdbid_ext}
\title{Javascript Generator Extended for plot db id}
\usage{
render_dup_plotdbid_ext(data)
}
\arguments{
\item{data}{The name of the data frame.}
}
\description{
Javascript code generator for render trait variables for Rhandsontable
}
\author{
alonso medina
}
