% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_tools.R
\name{render_duplicated_plot_dbid_ext}
\alias{render_duplicated_plot_dbid_ext}
\title{RJavascript generator for duplicated plot db id values}
\usage{
render_duplicated_plot_dbid_ext(dup_values)
}
\arguments{
\item{dup_values}{scale condition of the trait}
}
\description{
Function to generate Javascript code for duplicated plot db ids values in Rhansontable.
}
\author{
alonso medina
}
