% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_tools.R
\name{render_quantitative}
\alias{render_quantitative}
\title{RJavascript generator for quantitative traits in javascript conditions}
\usage{
render_quantitative(ll, ul)
}
\arguments{
\item{ll}{lower limit values of the trait}

\item{ul}{upper limmit values of the trait}
}
\description{
Function to generate Javascript code for cualitative trait validation in Rhansontable
}
\author{
omar benites
}

