% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_tools.R
\name{render_quantitative_ext}
\alias{render_quantitative_ext}
\title{RJavascript generator extended to render quantitative traits with javascript conditions and outlier detection}
\usage{
render_quantitative_ext(ll, ul, ol, ou)
}
\arguments{
\item{ll}{lower limit values of the trait}

\item{ul}{upper limmit values of the trait}

\item{ol}{lower outlier bound}

\item{ou}{upper outlier bound}
}
\description{
Function to generate Javascript code for quantitative trait validation and outlier detection
in Rhansontable
}
\author{
omar benites
}

