% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_tools.R
\name{render_trait}
\alias{render_trait}
\title{Javascript Generator for Trait Renders usigin Trai Dictionary from Crop Ontology}
\usage{
render_trait(trait, trait_dict, dsource = 1)
}
\arguments{
\item{trait}{trait}

\item{trait_dict}{trait dictionary}

\item{dsource}{source of the data \code{1}: Hidap,  \code{2}: FieldbookApp.}
}
\description{
Javascript code generator for render trait variables for Rhandsontable
}
\author{
omar benites
}
