% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_tools.R
\name{render_trait_ext}
\alias{render_trait_ext}
\title{Javascript Generator Extended for Trait Renders usigin Trai Dictionary from Crop Ontology}
\usage{
render_trait_ext(data, trait, trait_dict)
}
\arguments{
\item{data}{The name of the data frame.}

\item{trait}{trait}

\item{trait_dict}{trait dictionary}
}
\description{
Javascript code generator for render trait variables for Rhandsontable
}
\author{
omar benites
}

