% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saudpc}
\alias{saudpc}
\title{Scale for Area Under the Curve (Scale AUDPC)}
\usage{
saudpc(instn, audpc, reps, lb.ctrl)
}
\arguments{
\item{instn}{character value "INSTN" (The institucional number header)}

\item{audpc}{the audpc value}

\item{reps}{The number of repetitions}

\item{lb.ctrl}{The control genotype scale value of resistance for late blight trial.}
}
\description{
Function calculate relative days for Late Blight Resistance.
}
