% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subsample}
\alias{subsample}
\title{trial data using sub sampling for each observation}
\format{data frame}
\source{
International Potato Center, potato experimental data.
}
\usage{
subsample
}
\description{
This dataset contains data comec from experiment using sub sampling for each genotype.
}
\references{
This data is related to HiDAP fbcheck module.
}
