% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_stats.R
\name{summary_by_design}
\alias{summary_by_design}
\title{Summarize data by statistical design}
\usage{
summary_by_design(fieldbook, trait, genotype, factor, design = "RCBD",
  trait_dict)
}
\arguments{
\item{fieldbook}{A data.frame with the fieldbook data.}

\item{trait}{The column name of the trait into the fieldbook.}

\item{genotype}{The column name of the genotype}

\item{factor}{The column name of the factor}

\item{design}{The fieldbook design.}

\item{trait_dict}{A dataframe with the data dictionary based on crop ontologies.}
}
\value{
Return a data.frame summary based on statistical designs.
}
\description{
This function make a easy summary by clone (and treatment) based on statistical design
}

