% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_crop_dictionary.R
\docType{data}
\name{sweetpotato_yield}
\alias{sweetpotato_yield}
\title{Trait dictionary for Sweetpotato Yield Trial}
\format{A data frame with 39 columns and 53 rows.}
\source{
International Potato Center, sweetpotato experimental data.
}
\usage{
sweetpotato_yield
}
\description{
This data set contains a trait dicionary based on Crop Ontology template for
Healthy Tuber and Yield trail in Sweetpotato. This trait dictionary is developed at International
Potato Center.
}
\references{
This data dictionary is related to CO_331
}
