% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sweetpt}
\alias{sweetpt}
\alias{spg}
\title{Sample data of sweet potato}
\format{data frame}
\source{
International Potato Center, sweet potato experimental data.
}
\usage{
spg
}
\description{
This dataset contains sweet potato data
}
\references{
This data is related to HiDAP crop template
}
