% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_crop_dictionary.R
\docType{data}
\name{td_trait}
\alias{td_trait}
\title{Trait variables for Potato and Sweetpotato Trials}
\format{A data frame with 39 columns and 53 rows.}
\source{
International Potato Center, sweetpotato experimental data.
}
\usage{
td_trait
}
\description{
This data set contains a trait variables based on Crop Ontology template for
Potato and Sweetpotato. This trait dictionary is developed at International
Potato Center.
}
\references{
This data dictionary is related to CO_331
}
