% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_stats.R
\name{themode}
\alias{themode}
\title{Mode of the data.}
\usage{
themode(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector which contains the trait data.}

\item{na.rm}{Logical value to ommit NA values}
}
\description{
An statistical function to obtain the mode.
}
