% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_outlier_excel.R
\name{trait_outlier}
\alias{trait_outlier}
\title{An outlier values detector of sumarized fieldbook spreadsheets (Excel)}
\usage{
trait_outlier(file, sumsheet, trait)
}
\arguments{
\item{file}{The name of the excel file. It must be an .xlsx file}

\item{sumsheet}{The name of the excel sheet which contains the summary of your fieldbook .}

\item{trait}{The abbreviation of the trait used in fieldbooks.}
}
\description{
This function use openxlsx package to give conditional format for fieldbook spreadsheet
}

