% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_stats.R
\name{trait_summary}
\alias{trait_summary}
\title{Function to calculate summary statistics for fieldbook data from excel files}
\usage{
trait_summary(fieldbook = NULL, trait, groupfactors = NULL, na.rm = FALSE,
  trait_dict = NULL)
}
\arguments{
\item{fieldbook}{A data frame of the fieldbook}

\item{trait}{The name or position of the measured variable that will be summarized.}

\item{groupfactors}{A vector containing names of columns that contain grouping variables}

\item{na.rm}{A boolean that indicates whether to ignore NA's}

\item{trait_dict}{The data frame of the data dictionary for potato and sweetpotato}
}
\value{
A data frame with the count, mean and standard desviation
}
\description{
Function to calculate summary statistics for fieldbook data from excel files
}
\details{
This function is capable of divide the information in categorial or quantitative data based on a data dictionary for potato and sweepotato.
If it is categorical, returns the count #'and mode. And if it is quantitative, returns the count, media and standart desviation.
}
\author{
Omar Benites
}
\references{
Progress in developing a potato ontology for breeders. Reinhard Simon, Vilma Hualla, E. Salas, Rene Gomez, Raul Cordova and Stef de Haan.
Crop Ontology 2014.
}
\keyword{stats,}
\keyword{summary}

