% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_stats.R
\name{trait_summary_join}
\alias{trait_summary_join}
\title{Join all the summarized results of fieldbook's traits}
\usage{
trait_summary_join(fieldbook, trait, genotype = NA, factor = NA,
  design = "RCBD", trait_dict)
}
\arguments{
\item{fieldbook}{A data.frame with the fieldbook data.}

\item{trait}{The column name of the trait into the fieldbook.}

\item{genotype}{The column name of the genotype}

\item{factor}{The column name of the factor}

\item{design}{The fieldbook design.}

\item{trait_dict}{The trait dictionary on crop ontology format.}
}
\value{
An excel file with conditional format according trait conditions
}
\description{
This function sumarizes one or more traits acording to trait_dictionary. 
It show the mean (quantitative), mode (cualitative),
standard desviation and length of the genotypes.
}
