% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traittools_example.R
\name{traittoolsApp}
\alias{traittoolsApp}
\title{Run a data proccessing framework using the traittools functionality}
\usage{
traittoolsApp()
}
\description{
When the function is invoked a shiny app is started showing a very simple 
setup using traittools. A button summons the dialog box allowing the user to
navigate the R installation directory. First, you select your excel file, then 
traittools will show a spreadsheet identifying outlier and values out of range. Then press
on calculate for checking and applying formulas over the traits. Finally, you can download
your proccessing data using "Download" button with conditional formatting.
}
\seealso{
Other traitttols: \code{\link{traittoolsApp2}}
}
\concept{traitttols}
