% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_dates}
\alias{transform_dates}
\title{lapply transformation of dates using mutate_excel_date}
\usage{
transform_dates(mgt)
}
\arguments{
\item{mgt}{crop management sheet or table}
}
\description{
Transformation of dates using mgt
}
