% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xaudpc}
\alias{xaudpc}
\title{Get Relative Value for Area Under the Curve(AUDPC)}
\usage{
xaudpc(eval, dates, type)
}
\arguments{
\item{eval}{The trait evaluation}

\item{dates}{The dates of evaluation}

\item{type}{The type of trial}
}
\description{
Function calculate relative days for Late Blight Resistance.
}
