% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traittrender_rhansontable.R
\name{col_render_trait}
\alias{col_render_trait}
\title{Conditional Trait Render for Fieldbook Web Table}
\usage{
col_render_trait(fieldbook, trait, trait_dict, export = FALSE)
}
\arguments{
\item{fieldbook}{The fieldbook data}

\item{trait}{The abbreviation(s) of the trait(s) used in fieldbooks.}

\item{trait_dict}{The trait dictionary on crop ontology format.}

\item{export}{logical. If TRUE, using the right click, appear a context menu for downloading as \code{csv} file.
(paint with colours) the trait column to identify out of range values.}
}
\value{
An web table spreadsheet with conditional format according trait conditions
}
\description{
This function highlight all the values which are out of range in a fieldbook WebTable.
}

