% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_tools.R
\name{get_scale_trait}
\alias{get_scale_trait}
\title{Get a scale condition for quantitative or cualitative traits}
\usage{
get_scale_trait(trait, trait_dict)
}
\arguments{
\item{trait}{trait abreviation}

\item{trait_dict}{trait dictionary}
}
\description{
Function to get the scale of differents trait,
}
\examples{
\dontrun{
# The data
data(potato_yield)
#str(potato_yield)
# NTP: Number of tuber planted
get_scale_trait("NTP", potato_yield)
}
}
\author{
omar benites
}

