% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_tools.R
\name{get_trait_type}
\alias{get_trait_type}
\title{Get the type of trait into the trait dictionary}
\usage{
get_trait_type(trait, trait_dict)
}
\arguments{
\item{trait}{trait abreviation}

\item{trait_dict}{trait dictionary}
}
\value{
The type of trait
}
\description{
Get the type of trait into the trait dictionary
}
\examples{
\dontrun{
# The data
data(potato_yield)
#str(potato_yield)
# NTP: Number of tuber planted
get_trait_type("NTP", potato_yield)
}
}
\author{
omar benites
}

