% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_tools.R
\name{get_trait_units}
\alias{get_trait_units}
\title{Get the units of trait into the trait dictionary}
\usage{
get_trait_units(trait, trait_dict)
}
\arguments{
\item{trait}{trait abreviation}

\item{trait_dict}{trait dictionary}
}
\value{
The units of the trait in which has been measured
}
\description{
Get the units of trait into the trait dictionary
}
\examples{
\dontrun{
# The data
data(potato_yield)
#str(potato_yield)
# NTP: Number of tuber planted
get_trait_units("NTP", potato_yield)
}
}
\author{
omar benites
}
