% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_tools.R
\name{render_trait}
\alias{render_trait}
\title{Javascript Generator for Trait Renders usigin Trai Dictionary from Crop Ontology}
\usage{
render_trait(trait, trait_dict)
}
\arguments{
\item{trait}{trait}

\item{trait_dict}{trait dictionary}
}
\description{
Javascript code generator for render trait variables for Rhandsontable
}
\author{
omar benites
}

