% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_tools.R
\name{render_trait_ext}
\alias{render_trait_ext}
\title{Javascript Generator Extended for Trait Renders usigin Trai Dictionary from Crop Ontology}
\usage{
render_trait_ext(data, trait, trait_dict, dsource = 1)
}
\arguments{
\item{data}{The name of the data frame.}

\item{trait}{trait}

\item{trait_dict}{trait dictionary}

\item{dsource}{source of the data \code{1}: Hidap,  \code{2}: FieldbookApp.}
}
\description{
Javascript code generator for render trait variables for Rhandsontable
}
\author{
omar benites
}
